server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pyrebase4==4.7.1
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
beautifulsoup4==4.12.2
